refactor(queries): Rename create-memory #9
Better sort and visualization on the directory,
grouping the files by name.
