+++ src/memories/components/memory-form.connected.js
@@ -2,7 +2,7 @@ import { connect } from 'preact-redux'
 import { graphql } from 'react-apollo'
 import { reduxForm, formValueSelector, SubmissionError } from 'redux-form'
 
-import { createMemory } from '../queries'
+import { memoryCreate } from '../queries'
 import MemoryForm from './memory-form'
 
 const FORM = 'memoryForm'
@@ -11,13 +11,13 @@ const REQUIRED_FIELDS = [
   'ownerFirstName', 'ownerLastName', 'ownerEmail',
   'ownerCountry', 'victimName', 'victimBornAt', 'victimDeadAt',
   'victimCity', 'victimHistory', 'victimRememberText',
-  'victimPhoto', 'victimSilhouette', 'authorizedToSite'
+  'victimPhoto', 'victimSilhouette', 'authorizedToSite',
 ]
 
-const validate = values => {
+const validate = (values) => {
   const errors = {}
 
-  REQUIRED_FIELDS.map(fieldName => {
+  REQUIRED_FIELDS.map((fieldName) => {
     if (!values[fieldName]) {
       errors[fieldName] = 'Preenchimento obrigatório'
     }
@@ -26,27 +26,22 @@ const validate = values => {
   return errors
 }
 
-const mapStateToProps = state => {
+const mapStateToProps = (state) => {
   const selector = formValueSelector(FORM)
   return {
     ...selector(state, 'ownerFirstName', 'authorizedToSite'),
   }
 }
 
-export default graphql(createMemory, {
+export default graphql(memoryCreate, {
   props: ({ mutate }) => ({
-    onSubmit: values => {
-      return mutate({
-        variables: {...values }
-      })
-      .then(response => {
-        console.log(response)
-      })
-      .catch(error => {
-        throw new SubmissionError({ _error: '(500) Erro interno no servidor.' })
-      })
-    },
+    onSubmit: values =>
+      mutate({ variables: values })
+        .then((response) => { console.log(response) })
+        .catch(() => {
+          throw new SubmissionError({ _error: '(500) Erro interno no servidor.' })
+        }),
   }),
 })(connect(mapStateToProps)(
-  reduxForm({ form: FORM, validate })(MemoryForm)
+  reduxForm({ form: FORM, validate })(MemoryForm),
 ))
