+++ src/memories/queries/index.js
@@ -1,2 +1,2 @@
-export { default as createMemory } from './create-memory'
 export { default as memory } from './memory'
+export { default as memoryCreate } from './memory-create'
